minor cleanup and notes/todos

+++ os/src/arch/x86_64/net/ip/tcp.h
@@ -100,7 +100,7 @@ struct tcp_connection {
 
 	uint16_t window_size;
 
-	uint64_t new_data_avail;
+	uint8_t new_data_avail;
 	uint8_t received[IPV4_DATA_LEN];
 	uint8_t *received_top;
